% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sentiment.R
\name{count_sentiment}
\alias{count_sentiment}
\title{Get counts of the sentiments in your dataset}
\usage{
count_sentiment(data)
}
\arguments{
\item{data}{}
}
\value{
a table containing the counts of the sentiments in your data
}
\description{
Get counts of the sentiments in your dataset
}
\examples{
counted_sentiments <- count_sentiment(example_data$sentiments)
}
