% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiments.R
\name{get_sentiments}
\alias{get_sentiments}
\title{get_sentiments}
\usage{
get_sentiments(text, lexicon)
}
\arguments{
\item{lexicon}{}
}
\value{
a table containg the word tokens and their sentiment
}
\description{
get_sentiments
}
\examples{
example_strings <- c("I am really angry right now", "I am very very happy", "Rage and mixed feelings", "i lied")
example_sentiments <- get_sentiments(example_strings, "bing")
}
