% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sentiment_Vis.R
\name{sentiment_vis}
\alias{sentiment_vis}
\title{Sentiment_Vis}
\usage{
sentiment_vis(sentiments, words, index, vis_type)
}
\arguments{
\item{words}{}
}
\description{
Sentiment_Vis
}
\examples{
plot <- sentiment_vis(example_sentiments$sentiment, example_sentiments$indexes, vis_type = "sentiment_hist")
plot
}
